+++ public/phyfile.h
@@ -1,4 +1,4 @@
-//========= Copyright © 1996-2005, Valve Corporation, All rights reserved. ============//
+//========= Copyright Valve Corporation, All rights reserved. ============//
 //
 // Purpose: 
 //
@@ -17,7 +17,7 @@ typedef struct phyheader_s
 	int		size;
 	int		id;
 	int		solidCount;
-	long	checkSum;	// checksum of source .mdl file
+	int32	checkSum;	// checksum of source .mdl file
 } phyheader_t;
 
 #endif // PHYFILE_H
